+++ src/subcommand/info.rs
@@ -8,9 +8,9 @@ pub(crate) struct Info {
 
 #[derive(Serialize, Deserialize)]
 pub struct TransactionsOutput {
-  pub start: u64,
-  pub end: u64,
-  pub count: u64,
+  pub start: u32,
+  pub end: u32,
+  pub count: u32,
   pub elapsed: f64,
 }
